+++ packages/plugin/run/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @umac-js/run
 
+## 2.0.8
+
+### Patch Changes
+
+- add sleep and sleep-now methods
+
+- Updated dependencies []:
+  - @umac-js/utils@2.0.8
+
 ## 2.0.7
 
 ### Patch Changes