+++ packages/repo-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @umac-js/repo-config
 
+## 2.0.8
+
+### Patch Changes
+
+- add sleep and sleep-now methods
+
 ## 2.0.7
 
 ### Patch Changes